System: move Session class into its own namespace, add SessionFactory
